Change Flight:speedFraction type Float to Double

+++ MAS/MAS-ejb/src/java/ams/aps/entity/Flight.java
@@ -22,6 +22,7 @@ import javax.persistence.OneToMany;
  */
 @Entity
 public class Flight implements Serializable {
+
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,9 +32,9 @@ public class Flight implements Serializable {
     private Boolean completed;
     private Boolean scheduled;
     private Boolean deleted;
-    private Float speedFraction;
-    
-    @ManyToOne(cascade={CascadeType.ALL}, fetch=FetchType.EAGER)
+    private Double speedFraction;
+
+    @ManyToOne(cascade = {CascadeType.ALL}, fetch = FetchType.EAGER)
     private Route route;
 
     @OneToMany(cascade = {CascadeType.ALL}, fetch = FetchType.EAGER)
@@ -143,14 +144,14 @@ public class Flight implements Serializable {
     /**
      * @return the speedFraction
      */
-    public Float getSpeedFraction() {
+    public Double getSpeedFraction() {
         return speedFraction;
     }
 
     /**
      * @param speedFraction the speedFraction to set
      */
-    public void setSpeedFraction(Float speedFraction) {
+    public void setSpeedFraction(Double speedFraction) {
         this.speedFraction = speedFraction;
     }
 }
